Fixed NEAR error
The website was randomly getting stuck in a reloading-loop.
We found a bug in near, and this is a patch for it.
The bug is that on every call the used gas accumulates, and
after a certain number of calls, NEAR forced to authorize
the transaction. We fixed this by making two of the calls
asynchronic. Now, gas does not acumulate.